Harden and accelerate scoring and packing

+++ tmol/score/genbonded/potentials/genbonded_pose_score.hh
@@ -188,7 +188,7 @@ struct GenBondedRotamerScoreDispatch {
           TPack<Real, 2, D>,          // V_t
           TPack<Vec<Real, 3>, 2, D>,  // dV_dx_t
           TPack<Int, 2, D>,           // dispatch_indices_t
-          TPack<Int, 1, D>,           // n_output_intxns_for_rot_conn_offset
+          TPack<int64_t, 1, D>,       // n_output_intxns_for_rot_conn_offset
           TPack<Int, 1, D>            // rotconn_for_output_intxn
           >;
 
@@ -223,7 +223,7 @@ struct GenBondedRotamerScoreDispatch {
       TView<Vec<Real, 5>, 1, D> gen_inter_improper_hash_values,
 
       TView<Int, 2, D> dispatch_indices,
-      TView<Int, 1, D> n_output_intxns_for_rot_conn_offset,
+      TView<int64_t, 1, D> n_output_intxns_for_rot_conn_offset,
       TView<Int, 1, D> rotconn_for_output_intxn,
       TView<Real, 2, D> dTdV) -> TPack<Vec<Real, 3>, 2, D>;
 };
